packages/python/adapter/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otlp-stdout-adapter"
version = "0.1.0"
description = "A custom requests HTTPAdapter that serializes spans to stdout for OpenTelemetry OTLP exporters"
readme = "README.md"
authors = [{name = "Alessandro Bologna", email = "alessandro.bologna@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
    "Typing :: Typed"
]
keywords = ["opentelemetry", "otlp", "stdout", "lambda", "aws"]
dependencies = [
    "opentelemetry-api~=1.27.0",
    "opentelemetry-sdk~=1.27.0",
    "opentelemetry-exporter-otlp-proto-http~=1.27.0",
    "opentelemetry-semantic-conventions~=0.48b0",
    "requests>=2.31.0,<3.0.0",
]
requires-python = ">=3.12"

[project.urls]
Homepage = "https://github.com/dev7a/lambda-otlp-forwarder/tree/main/packages/python/adapter"
Repository = "https://github.com/dev7a/lambda-otlp-forwarder/tree/main/packages/python/adapter"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}
packages = ["otlp_stdout_adapter"]

[tool.setuptools.exclude-package-data]
"*" = ["tests/*", "examples/*"]

[tool.poetry]
name = "otlp-stdout-adapter"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
    "pylint>=2.17.0"
]
